@@ -19,10 +19,10 @@
     "node-modules": "^1.0.1",
     "nodemailer": "^6.7.5",
     "phosphor-react": "^1.4.1",
-    "react": "^18.1.0",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.4.0",
     "react-bootstrap-icons": "^1.8.2",
-    "react-dom": "^18.1.0",
+    "react-dom": "^18.3.1",
     "react-mailchimp-subscribe": "^2.1.3",
     "react-multi-carousel": "^2.8.1",
     "react-on-screen": "^2.1.1",
