@@ -1,12 +1,11 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import Routes from "./Routes";
+import AppRoutes from "./AppRoutes";
 
 function App() {
   return (
     <div>
-      <Routes />
+      <AppRoutes />
     </div>
   );
 }
